Change int to float

+++ 0x00-python_variable_annotations/0-add.py
@@ -4,15 +4,15 @@ This is a module that provides a function for adding two numbers.
 """
 
 
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> float:
     """
     This function adds two numbers and returns the result.
 
     Parameters:
-    a (int): The first number.
-    b (int): The second number.
+    a (float): The first number.
+    b (float): The second number.
 
     Returns:
-    int: The sum of a and b.
+    float: The sum of a and b.
     """
     return a + b
